# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gearbox API",
        "description": "REST API untuk Sistem Booking & Manajemen Layanan Bengkel Mobil",
        "contact": {
            "email": "admin@gearbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get kendaraan milik user yang sedang login",
                "operationId": "043b3ce51f006f946511103cfe1c2ca1",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List kendaraan milik user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "plate_number": {
                                                        "type": "string",
                                                        "example": "B 1234 ABC"
                                                    },
                                                    "brand": {
                                                        "type": "string",
                                                        "example": "Toyota"
                                                    },
                                                    "model": {
                                                        "type": "string",
                                                        "example": "Avanza"
                                                    },
                                                    "year": {
                                                        "type": "integer",
                                                        "example": 2020
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "mobil"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Tambah kendaraan baru",
                "operationId": "74e9ae3e5f1ee2f69f3a177820d204b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plate_number",
                                    "brand",
                                    "model",
                                    "year",
                                    "type"
                                ],
                                "properties": {
                                    "plate_number": {
                                        "description": "Plat nomor unik kendaraan",
                                        "type": "string",
                                        "example": "B 1234 ABC"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Avanza"
                                    },
                                    "year": {
                                        "description": "Tahun kendaraan (1990 - sekarang)",
                                        "type": "integer",
                                        "example": 2020
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "mobil",
                                        "enum": [
                                            "motor",
                                            "mobil"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kendaraan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kendaraan berhasil ditambahkan"
                                        },
                                        "vehicle": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal atau plat nomor sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicles/{vehicle}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get detail kendaraan milik user",
                "operationId": "2d7291bedeeb73a49905966effb2dee4",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kendaraan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "plate_number": {
                                            "type": "string",
                                            "example": "B 1234 ABC"
                                        },
                                        "brand": {
                                            "type": "string",
                                            "example": "Toyota"
                                        },
                                        "model": {
                                            "type": "string",
                                            "example": "Avanza"
                                        },
                                        "year": {
                                            "type": "integer",
                                            "example": 2020
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "mobil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Update kendaraan milik user",
                "operationId": "82a1b2a717b16eb7e08a86288a7f8093",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plate_number": {
                                        "type": "string",
                                        "example": "B 5678 XYZ"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Honda"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Jazz"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "motor",
                                            "mobil"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kendaraan berhasil diupdate"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Hapus kendaraan milik user",
                "operationId": "495952c6402028fab9f82c080af676ad",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "ID kendaraan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kendaraan berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Kendaraan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token dari response login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}